Remove status indicators on the Tree dropdown items text.
https://github.com/silverstripe/silverstripe-framework/issues/5702

I'd say just remove the status indicators. Pointer for how to fix: Make a change to a page and only save it without publication, which should cause a "modified" status in the tree. Go to admin/pages/add and choose "under another page". The _tree.scss styles contain a tree-status-icon mixin which is responsible for this styling. That's not replicated in TreeDropdownField.scss