FIX amend preview URL correctly, more flexibly
When using either split mode or preview mode while
working with content in the CMS, the preview pane
would simply tack ?CMSPreview=1 on to the end of
whatever URL it is given. This is not ideal when
working with e.g. Elemental, as there is a fragment
for jumping to the relevant content in the preview,
so we end us with some/page?Stage=stage#someBlock&CMSPreview=1
which counts as neither a CMSpreview query, nor a vaild (existing)
fragment identifier. So I've patched it to correctly amend to
e.g. some/page?Stage=stage&CMSpreview=1#someBlock