ENHANCEMENT Register fieldHolder HOCs with injector
rather than having their default export be a higher
order component of the field proper. This allows
for more intuitive use of the Injector API, particularly
around the use of transforms. An end user can now modify
or transform the FormField proper, as opposed to a Higher
Order Component in which it sits (which changes the
interface one expects).