NEW Add CMS community help menu to cms-menu
This commit removes the "Help" menu item and adds a userhelp, docs, community and feedback help menu to the south toolbar.
When the cms-menu is closed the SS logo shows in the cms-menu header and the new help menu is removed.
Includes changes to the cms version badge and content that is displayed.
The help links are also configurable making it easy for agencies to remove or add their own.
Also addresses a fix for the mobile menu toggle functionality.
Co-authored by: Scott Hutchinson https://github.com/scott1702