FIX / TreeDropdownField empty value to be null
The PR removes the legacy implementation of empty values for the component.
It used to be a string "unchanged" for multiselect field and a 0 for a single value.
However, we should avoid making such assumptions so that option id may be a string
in the future (e.g. uuid) as well as the word "unchanged" must not have any special meaning.

Related github issues:
 - silverstripe-assets/#176
 - silverstripe-framework/#8596
 - silverstripe-framework/#8332
 - silverstripe-framework/#8672