FIX Scrolling out of auto-selected edit mode not switches back to split mode
Previously resizing the window in split mode from a size which supported split mode
into one that does not would switch to edit mode, but it would not switch back to
split mode when resizing back up again. This patch fixes that by storing an extra
"state is locked" value to ensure that when you manually select "Edit" or "Preview"
mode, it will not switch back to split mode when you resize, but also ensures that
it can switch between split and edit on resize automatically otherwise.