FIX re-enable ability to insert anchors
SilverStripe 3 had the ability to insert anchors into content areas
within the CMS. SilverStripe 4 however has separated the ideas of the
admin area and the CMS - however the ability for an editor to insert
an anchor is not directly related to pages. To this end the anchor
plugin for TinyMCE should be enabled here, as opposed to globally
for all rich text editors, e.g. enabling for unauthenticated users
adding blog comments shouldn't be done.

This has the knock on affect for the CMS module, where the "link to
an anchor on another page" plugin is also enabled. Since the release
of SilverStripe 4 it has been an option to create such links, but an
editor could not also create the anchors to link to. Now they will be
able to do such.