FIX refresh tinyMCE after it is moved in the DOM
When a TinyMCE instance is moved in the DOM, most especially if it is
moved down the DOM tree, the content will disappear and many of the
interactions (toolbar buttons) will cease functioning as intended.

The only way to 'refresh' an instance seems to be to remove it then
re-add it, so this commit introduces this functionality to the repaint
method in entwine, meaning the editor will continue to use the same
config it was first instantiated with.