Defensive change
I'm not sure why, but on page load, the reducer.store may not be available. This change ensures the page loads regardless of the reducer. i know it's probably not a good idea to log stuff to the console, but i have no idea how to fix the missing reducer

See #1163