FIX Unload previous preview when new URL is not available
Some pagetypes (and other previewable objects) may not have any
available preview states. A primary example of this is RedirectorPage.
When switching the editor to one of these pagetypes, the prior preview
was retained, though covered up by a broken overlay. This change, along
with the removal of the broken overlay, means that a simple 'No preview
available' message will be successfully displayed in these situations.

This accompanies a template fix in the CMS module.