ENH: Allow use of the preview panel in any LeftAndMain subclass.
For some reason, the preview panel template which is referred to in LeftAndMain::PreviewPanel() only exists in the cms module for the CMSMain controller.
All of the other code for using a CMS preview is available for any arbitrary LeftAndMain subclass and is available in this module. There should be no dependency from this module on the cms module to allow previews, and previews should work for all LeftAndMain subclasses in this module (e.g. ModelAdmin).

Relies on silverstripe/silverstripe-framework#10112