FIX Restore correct widths to main CMS panel.
The introduction of a new parent to .cms-main resulted in some of the
CSS and JS that handles the width of the main panel being incorrect.
Visually it appeared that the left and right panels were too narrow -
but in reality that's just because the main was was too wide.