ENH Allow adding filter prefix to filter props
In CMSMain and gridfield search filters, the `filters` prop is passed in
from HTML data attributes, which already include the prefix.

In AssetAdmin, the filters NEVER include the prefix but don't need to,
because they don't use SearchContext.

For CampaignAdmin, since we're not re-rendering the searchbox serverside
every time we search (or ever) we can't rely on the CMSMain way of doing
things - but since we're using SearchContext we DO need the prefix
passed into the prop (even though we don't want it in the search POST
request).